refactor(desktop): share browser cookie result types
