+++ apps/desktop/src/preview/BrowserImport/Sources.ts
@@ -299,7 +299,7 @@ export const sourcePathContext = Effect.gen(function* () {
 const entryExists = Effect.fnUntraced(function* (path: string) {
   const fileSystem = yield* FileSystem.FileSystem;
   return yield* fileSystem.stat(path).pipe(
-    Effect.catchCause(() => fileSystem.readLink(path)),
+    Effect.catch(() => fileSystem.readLink(path)),
     Effect.as(true),
     Effect.orElseSucceed(() => false),
   );
